# core/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: NeuroGate API
  description: |
    Enterprise AI Gateway API providing intelligent LLM routing, PII protection,
    semantic caching, and multi-provider orchestration.

    ## Features
    - **Multi-Provider Routing**: Intelligent routing across OpenAI, Anthropic, Google, AWS Bedrock, and Azure
    - **PII Protection**: Automatic detection and redaction of sensitive data (NeuroGuard)
    - **Semantic Caching**: 4-tier caching with semantic similarity matching
    - **Cost Tracking**: Real-time usage and cost analytics
    - **Agent Operations**: Trace and monitor AI agent workflows

    ## Authentication
    Use API key authentication via the `Authorization` header:
    ```
    Authorization: Bearer your-api-key
    ```

  version: 1.0.0
  contact:
    name: NeuroGate Team
    url: https://neurogate.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.neurogate.dev
    description: Production
...
paths:
...
  /v1/models/{modelId}:
    get:
      tags: [Models]
      summary: Get model details
      description: Get detailed information about a specific model
      operationId: getModel
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelInfo'
        '404':
          description: Model not found
...
  /api/v1/synapse/prompts/{promptName}/workflow:
    get:
      tags: [Synapse]
      summary: Get prompt workflow
      operationId: getPromptWorkflow
      parameters:
        - name: promptName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptWorkflow'
  /api/v1/synapse/deploy:
    post:
      tags: [Synapse]
      summary: Deploy prompt
      operationId: deployPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployRequest'
      responses:
        '200':
          description: Prompt deployed
  /api/v1/synapse/play:
    post:
      tags: [Synapse]
      summary: Playground execution
      operationId: playPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayRequest'
      responses:
        '200':
          description: Execution completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
...
components:
  schemas:
...
    ChatResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          default: chat.completion
        model:
          type: string
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Choice'
        usage:
          $ref: '#/components/schemas/Usage'
        cache_hit:
          type: boolean
        route:
          type: string
    Message:
      type: object
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string
        name:
          type: string
        tool_calls:
          type: array
          items:
            type: object
    Choice:
      type: object
      properties:
        index:
          type: integer
        message:
          $ref: '#/components/schemas/Message'
        finish_reason:
          type: string
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer

    # ==================== Models Schemas ====================
...
    ModelInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        input_cost_per_1k:
          type: number
        output_cost_per_1k:
          type: number
        context_window:
          type: integer
        max_output_tokens:
          type: integer
        capabilities:
          type: array
          items:
            type: string
        available:
          type: boolean
...
    PromptWorkflow:
      type: object
      properties:
        prompt_name:
          type: string
        production_version:
          type: string
        staging_version:
          type: string
    DeployRequest:
      type: object
      properties:
        prompt_name:
          type: string
        version_id:
          type: string
        environment:
          type: string
        user:
          type: string
    PlayRequest:
      type: object
      properties:
        prompt_content:
          type: string
        variables:
          type: object
        model:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []